Implementation guide and examples updated
