--- packages/terra-framework-docs/src/terra-dev-site/doc/compact-interactive-list/Examples/ResponsiveColumnsMaxWidth.jsx
@@ -1,82 +1,0 @@
-import React from 'react';
-import CompactInteractiveList, { alignTypes } from 'terra-compact-interactive-list';
-import { IconDocuments, IconFeaturedOutline, IconImage } from 'terra-icon';
-
-const rows = [
-  {
-    id: 'row_1',
-    cells: [
-      { content: <IconDocuments a11yLabel="Documents" height="1.5em" width="1.5em" /> },
-      { content: 'Discern Care Set (1)' },
-      { content: <IconFeaturedOutline a11yLabel="Featured" height="1.5em" width="1.5em" /> },
-    ],
-  },
-  {
-    id: 'row_2',
-    cells: [
-      { content: <IconImage a11yLabel="Picture" height="1.5em" width="1.5em" /> },
-      { content: 'Initial observation Care/Day High Severity 99220 (2)' },
-      { content: <IconFeaturedOutline a11yLabel="Featured" height="1.5em" width="1.5em" /> },
-    ],
-  },
-  {
-    id: 'row_3',
-    cells: [
-      { content: <IconImage a11yLabel="Picture" height="1.5em" width="1.5em" /> },
-      { content: 'Arterial Sheath Care (3)' },
-      { content: <IconFeaturedOutline a11yLabel="Featured" height="1.5em" width="1.5em" /> },
-    ],
-  },
-  {
-    id: 'row_4',
-    cells: [
-      { content: ' ' },
-      { content: 'Sbsq Observation Care/Day High Severity 99226 (4)' },
-      { content: ' ' },
-    ],
-  },
-  {
-    id: 'row_5',
-    cells: [
-      { content: <IconImage a11yLabel="Picture" height="1.5em" width="1.5em" /> },
-      { content: 'Arterial Sheath Care (5)' },
-      { content: <IconFeaturedOutline a11yLabel="Featured" height="1.5em" width="1.5em" /> },
-    ],
-  },
-];
-
-const cols = [
-  {
-    id: 'Column-0',
-    displayName: 'Icon',
-    width: '60px', // there is no widthUnit prop set in the component, so it will default to px
-    align: alignTypes.CENTER,
-  },
-  {
-    id: 'Column-1',
-    displayName: 'Service name',
-    width: '200px', // will be disregarded because of flexGrow prop
-    flexGrow: true, // flexGrow prop has to be set here, as with width set, without that prop the column would not grow.
-    maximumWidth: '300px',
-    minimumWidth: '100px',
-  },
-  {
-    id: 'Column-2',
-    displayName: 'Featured',
-    width: '60px', // there is no widthUnit prop set in the component, so it will default to px
-    align: alignTypes.CENTER,
-  },
-];
-
-const ResponsiveColumnsMaxWidth = () => (
-  <CompactInteractiveList
-    id="with-breakpoints"
-    rows={rows}
-    columns={cols}
-    numberOfColumns={2}
-    columnMinimumWidth="234px"
-    rowHeaderIndex={1}
-  />
-);
-
-export default ResponsiveColumnsMaxWidth;
